Actually make tests run
A couple errors here:

- The index in argv we were accessing to get the file name was
wrong.
- The conversion of AVAILABLE_MEMORY to u16 was causing the
result to be zero (makes sense! AVAILABLE_MEMORY > 2^16), so
the CPU was never starting. Instead of converting the constant
to the type of the register, I decided to cast the register to
the type of the constant, which is a safer conversion since it
goes to an integer with more bits, and not to one with less.

+++ mos6502cpu/src/mos6502cpu.rs
@@ -380,7 +380,7 @@ impl Cpu<u8, Mos6502Instruction, CpuError> for Mos6502Cpu {
     }
 
     fn is_done(&self) -> bool {
-        self.registers.pc >= AVAILABLE_MEMORY as u16
+        self.registers.pc as usize >= AVAILABLE_MEMORY
     }
 
     fn increase_pc(&mut self, steps: u8) {
